requirements.txt
semantic-kernel>=1.34.0
python-dotenv>=1.0.0
azure-identity>=1.17.0
numpy>=1.26.0
aiohttp>=3.9.0
pydantic>=1.10.0
pydantic-settings>=2.7.0
rich>=13.0.0
colorlog>=6.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pyyaml>=6.0.0

# Additional LLM providers
openai>=1.6.7